pyproject.toml
[tool.poetry]
name = "sp-devsup-vault-expiration-monitoring"
version = "0.1.0"
description = "Prometheus exporter to monitor custom metadata for KV2 secrets for (self-imposed) expiration."
authors = ["Eugene Davis <eugene.davis@tomtom.com>"]

packages = [
    { include = "expiration_monitor" },
]

[tool.poetry.dependencies]
python = "^3.8"
hvac = "0.11.2"
prometheus-client = "0.14.1"
requests = "2.27.1"
PyYAML = "6.0"

[tool.poetry.dev-dependencies]
black = "22.3.0"
mypy = "0.942"
pylint = "2.13.7"
pytest = "7.1.1"
pytest-cov = "3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 200


[tool.mypy]

[[tool.mypy.overrides]]
ignore_missing_imports = true


[tool.pylint]

[tool.pylist.MASTER]
jobs = 0
limit-inference-results = 100

[tool.pylint.'MESSAGES CONTROL']
disable = [
   "print-statement",
   "parameter-unpacking",
   "unpacking-in-except",
   "old-raise-syntax",
   "backtick",
   "long-suffix",
   "old-ne-operator",
   "old-octal-literal",
   "import-star-module-level",
   "non-ascii-bytes-literal",
   "raw-checker-failed",
   "bad-inline-option",
   "locally-disabled",
   "file-ignored",
   "suppressed-message",
   "useless-suppression",
   "deprecated-pragma",
   "use-symbolic-message-instead",
   "apply-builtin",
   "basestring-builtin",
   "buffer-builtin",
   "cmp-builtin",
   "coerce-builtin",
   "execfile-builtin",
   "file-builtin",
   "long-builtin",
   "raw_input-builtin",
   "reduce-builtin",
   "standarderror-builtin",
   "unicode-builtin",
   "xrange-builtin",
   "coerce-method",
   "delslice-method",
   "getslice-method",
   "setslice-method",
   "no-absolute-import",
   "old-division",
   "dict-iter-method",
   "dict-view-method",
   "next-method-called",
   "metaclass-assignment",
   "indexing-exception",
   "raising-string",
   "reload-builtin",
   "oct-method",
   "hex-method",
   "nonzero-method",
   "cmp-method",
   "input-builtin",
   "round-builtin",
   "intern-builtin",
   "unichr-builtin",
   "map-builtin-not-iterating",
   "zip-builtin-not-iterating",
   "range-builtin-not-iterating",
   "filter-builtin-not-iterating",
   "using-cmp-argument",
   "eq-without-hash",
   "div-method",
   "idiv-method",
   "rdiv-method",
   "exception-message-attribute",
   "invalid-str-codec",
   "sys-max-int",
   "bad-python3-import",
   "deprecated-string-function",
   "deprecated-str-translate-call",
   "deprecated-itertools-function",
   "deprecated-types-field",
   "next-method-defined",
   "dict-items-not-iterating",
   "dict-keys-not-iterating",
   "dict-values-not-iterating",
   "deprecated-operator-function",
   "deprecated-urllib-function",
   "xreadlines-attribute",
   "deprecated-sys-function",
   "exception-escape",
   "comprehension-escape",
    "bad-continuation"
]

[tool.pylint.REFACTORING]
max-nested-blocks = 5

[tool.pylint.FORMAT]
max-line-length = 200
max-module-lines = 1000

[tools.pylint.SIMILARITIES]
min-similarity-lines = 4

[tool.pytest.ini_options]
minversion = "6.2.5"
testpaths = ["tests"]
